Stream company + watchlist pages via Suspense
Apply the same streaming pattern from the explore page:
- Split heavy data fetching into async server components
  (CompanyContent, WatchlistContent) wrapped in Suspense
- Page shells resolve instantly, streaming skeleton fallbacks
- Add shared skeleton components and loading.tsx for route transitions
- Fix searchParams types to handle string[] (same as explore page)
