+++ apps/web/app/[lang]/(app)/[userSlug]/[watchlistSlug]/loading.tsx
@@ -1,0 +1,5 @@
+import { WatchlistSkeleton } from "@/components/search/watchlist-skeleton";
+
+export default function Loading() {
+  return <WatchlistSkeleton />;
+}
