+++ apps/web/app/[lang]/(app)/company/[slug]/loading.tsx
@@ -1,0 +1,5 @@
+import { CompanySkeleton } from "@/components/search/company-skeleton";
+
+export default function Loading() {
+  return <CompanySkeleton />;
+}
